fix: fix na ProdutoRepository e na ProdutoJpaController

+++ dev/back-end/BackLaizeTech/BackLaizeTech/src/main/kotlin/laize_tech/back/repository/ProdutoRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 
 
-interface ProdutoRepository : JpaRepository<Produto, Long> {
+interface ProdutoRepository : JpaRepository<Produto, Int> {
 
     @Query(
         value = """
